fix: address code review feedback
- Move test_create_api_rejects_no_image_source from test_models.py to
  test_api.py (it's an API test, not a model test)
- Migrate _get_auth_header in test_auth.py to auth_header_for_user
- Remove shadow client fixture and unused Client import from test_auth.py
