+++ apps/pictograms/tests/test_models.py
@@ -73,15 +73,3 @@ class TestPictogramValidation:
 
         p = Pictogram.objects.create(name="File Only", image=make_test_image())
         assert p.pk is not None
-
-    def test_create_api_rejects_no_image_source(self, client, owner, org):
-        from conftest import auth_header_for_user
-
-        headers = auth_header_for_user(owner)
-        response = client.post(
-            "/api/v1/pictograms",
-            data={"name": "No Image", "image_url": "", "organization_id": org.id, "generate_sound": False},
-            content_type="application/json",
-            **headers,
-        )
-        assert response.status_code == 422
